rentAccess/requirements.txt
amqp==5.0.5
anyjson==0.3.3
asgiref==3.3.1
astroid==2.5.1
atomicwrites==1.4.0
attrs==20.3.0
billiard==3.6.3.0
boto3==1.17.17
botocore==1.20.17
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
decorators==2.0.7
dj-database-url==0.5.0
dj-rest-auth==2.1.3
Django==3.1.7
django-autoslug==1.9.8
django-cities-light==3.8.1
django-cors-headers==3.7.0
django-countries==7.0
django-db-logger==0.1.10
django-debug-toolbar==3.2
django-environ==0.4.5
django-extensions==3.1.1
django-filter==2.4.0
django-phone-field==1.8.1
django-redis==4.12.1
django-rest-framework-mongoengine==3.4.1
django-rest-swagger==2.2.0
django-searchable-encrypted-fields==0.1.9
django-settings-json==1.0.2
django-storages==1.11.1
django-template-preview==0.1.5
django-timezone-field==4.1.1
django-watchman==1.2.0
djangorestframework==3.12.2
djangorestframework-api-key==2.0.0
djangorestframework-simplejwt==4.6.0
dodgy==0.2.1
drf-yasg==1.20.0
factory-boy==3.2.0
Faker==6.5.0
flake8==3.8.4
flake8-polyfill==1.0.2
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
iniconfig==1.1.1
isort==5.7.0
itypes==1.2.0
Jinja2==2.11.3
jmespath==0.10.0
jsonschema==3.2.0
kombu==5.0.2
lazy-object-proxy==1.5.2
MarkupSafe==1.1.1
mccabe==0.6.1
mixins==0.1.4
mongoengine==0.22.1
numpy==1.20.1
openapi-codec==1.3.2
packaging==20.9
pandas==1.2.3
pendulum==2.1.2
Pillow==8.1.0
pluggy==0.13.1
progressbar2==3.53.1
prompt-toolkit==3.0.16
psycopg2-binary==2.8.6
py==1.10.0
pycodestyle==2.6.0
pycryptodome==3.10.1
pydocstyle==5.1.1
pyflakes==2.2.0
PyJWT==2.0.1
pylint==2.7.2
pylint-celery==0.3
pylint-django==2.4.2
pylint-flask==0.6
pylint-plugin-utils==0.6
pymongo==3.11.3
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.2
pytest-cov==2.11.1
pytest-django==4.1.0
python-dateutil==2.8.1
python-utils==2.5.6
pytz==2021.1
pytzdata==2020.1
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
requirements-detector==0.7
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.4
setoptconf==0.2.0
simplejson==3.17.2
six==1.15.0
snowballstemmer==2.1.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
Unidecode==1.2.0
uritemplate==3.0.1
urllib3==1.26.3
vine==5.0.0
vobject==0.9.6.1
wcwidth==0.2.5
wrapt==1.12.1
